Fix project name discovery
One folder that will always be there is the `.git` folder (as we have
cloned git repositories).  The second one could be `test` or `tests`,
where the tests reside usually.  This could be error prone in cases
where tests and code are in the same folder, but this seems uncommon, so
this work-around seems legitimate here.  Keep this in mind and maybe
change it if necessary.