Implement run-time limit for test executions
This enables the library to kill long-running test executions after a
hard time limit.  Use the optional parameter `time_limit` of the
`Runner` constructor to set a value larger than 0 as a time limit in
seconds.  No value, or the value 0, will be treated as infinite run time
for the runner.

This feature uses BenchExec, cf. https://github.com/sosy-lab/benchexec